+++ activator/src/processor.rs
@@ -770,6 +770,8 @@ mod tests {
             side_z_iface_name: "Ethernet1".to_string(),
             link_health: doublezero_serviceability::state::link::LinkHealth::Pending,
             desired_status: doublezero_serviceability::state::link::LinkDesiredStatus::Activated,
+            link_topologies: vec![],
+            link_flags: 0,
         };
 
         let mut existing_links: HashMap<Pubkey, Link> = HashMap::new();
@@ -804,6 +806,8 @@ mod tests {
             side_z_iface_name: "Ethernet3".to_string(),
             link_health: doublezero_serviceability::state::link::LinkHealth::Pending,
             desired_status: doublezero_serviceability::state::link::LinkDesiredStatus::Activated,
+            link_topologies: vec![],
+            link_flags: 0,
         };
 
         let new_link_cloned = new_link.clone();