+++ smartcontract/cli/src/user/get.rs
@@ -197,6 +197,7 @@ mod tests {
             metro_routing: false,
             route_liveness: false,
             billing: TenantBillingConfig::default(),
+            include_topologies: vec![],
         };
 
         let device_pubkey = Pubkey::new_unique();
@@ -389,6 +390,7 @@ mod tests {
             metro_routing: false,
             route_liveness: false,
             billing: TenantBillingConfig::default(),
+            include_topologies: vec![],
         };
 
         let device_pubkey = Pubkey::new_unique();